Split out NewClusterModal to own file
