+++ src/pages/workspaces/workspace/notebooks/NotebookLauncher.js
@@ -3,9 +3,9 @@ import { forwardRef, Fragment, useRef, useState } from 'react'
 import { div, h, iframe } from 'react-hyperscript-helpers'
 import * as breadcrumbs from 'src/components/breadcrumbs'
 import { requesterPaysWrapper, withRequesterPaysHandler } from 'src/components/bucket-utils'
-import { NewClusterModal } from 'src/components/ClusterManager'
 import { ButtonOutline, ButtonPrimary, Link } from 'src/components/common'
 import { spinner } from 'src/components/icons'
+import { NewClusterModal } from 'src/components/NewClusterModal'
 import { notify } from 'src/components/Notifications'
 import { Ajax, useCancellation } from 'src/libs/ajax'
 import { withErrorReporting } from 'src/libs/error'
